1.7.6: Hot fix - Using double instead of int for scores - Pinging the database to make sure it's connected - Allowed < 6 games stats to be counted - Added Deprecated games to the game list - Changed colors in Games.java to be ColorType instead of TextColor - For technical reasons, when initiated, PlayerStats don't have any scores. - Fixed a typo - Fixed games not showing in tab completion in /stats - Lowered points in tab completion to 10000 - Updated increment in tab completion to 0.5 instead of 1

+++ src/main/java/fr/thefox580/theevent5802/utils/Score.java
@@ -5,7 +5,7 @@ import java.util.Objects;
 public class Score {
 
     private Game game;
-    private int points = 0;
+    private double points = 0;
 
     public Score(Game game){
         this.game = game;
@@ -19,11 +19,11 @@ public class Score {
         this.game = game;
     }
 
-    public int getPoints() {
+    public double getPoints() {
         return this.points;
     }
 
-    public void setPoints(int points) {
+    public void setPoints(double points) {
         this.points = points;
     }
